requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
gradio_client>=2.5.0
httpx==0.27.2
Pillow==11.0.0
